+++ Warkey.Core/Warkey.Core.csproj
@@ -42,10 +42,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Services\ForegroundWindowService.cs" />
     <Compile Include="Services\GameVideoSettingsService.cs" />
     <Compile Include="MainService.cs" />
-    <Compile Include="Services\MainWindowHandleService.cs" />
     <Compile Include="Services\UpdateService.cs" />
     <Compile Include="Settings.cs" />
   </ItemGroup>
