+++ Warkey.Infrastructure/Game/Settings.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace WarkeyNETIII.Infrastructure.External
+namespace WarkeyNETIII.Infrastructure.Game
 {
     public class GameSettings
     {
