+++ Warkey.Infrastructure/Keyboard/Detector.cs
@@ -55,14 +55,14 @@ namespace Warkey.Infrastructure.Keyboard
 
         private delegate IntPtr LowLevelKeyboardProc(int nCode, IntPtr wParam, IntPtr lParam);
 
-        private IntPtr HookCallback(int nCode, IntPtr wParam, IntPtr lParam)
+        private IntPtr HookCallback(Action<bool> preprocessCheck, int nCode, IntPtr wParam, IntPtr lParam)
         {
             // iswar3foreground have to put here
             // because if not in foreground no need to do anything
             // else need to have pointer to mainservice iswar3foreground
             // and pass in here when initialize
             // which is not possible with C#
-            if (MainService.IsWar3Foreground && !ChatboxService.IsChatting && nCode >= 0)
+            if (preprocessCheck && !ChatboxService.IsChatting && nCode >= 0)
             {
                 int vkCode = Marshal.ReadInt32(lParam);
