+++ Warkey.Infrastructure/Warkey.Infrastructure.csproj
@@ -46,13 +46,14 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ClientSettings.cs" />
+    <Compile Include="Settings.cs" />
     <Compile Include="Game\Settings.cs" />
+    <Compile Include="Game\Window.cs" />
     <Compile Include="Keyboard\Detector.cs" />
     <Compile Include="Keyboard\Sender.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Devices\Monitor.cs" />
-    <Compile Include="Game\TKoKSaveFiles.cs" />
+    <Compile Include="Game\Saves.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Warkey.Shared\Warkey.Shared.csproj">
